src/requirements.txt
discord.py
colorama
asyncio
requests
Pillow
discord.py-pagination
asyncpraw
pyongc
twirl
selenium